migrate 'guardAccess' calls to 'backend', rename 'package' to 'core'
