+++ apps/core/package.json
