+++ apps/core/scripts/generate-langs-json.js
