+++ apps/core/scripts/generate-theme-doc.js
