+++ apps/core/src/common/config.js
