+++ apps/core/src/common/constants.js
