+++ apps/core/src/common/fmt.js
