+++ apps/core/src/common/html.js
