+++ apps/core/src/common/icons.js
