+++ apps/core/src/common/languageColors.json
