+++ apps/core/src/common/log.js
