+++ apps/core/src/common/render.js
