+++ apps/core/src/fetchers/stats.js
