+++ apps/core/src/fetchers/types.d.ts
