+++ apps/core/src/translations.js
