+++ apps/core/tsconfig.json
