--- apps/package/api/index.js
@@ -1,197 +1,0 @@
-// @ts-check
-
-import { renderStatsCard } from "../src/cards/stats.js";
-import { guardAccess } from "../src/common/access.js";
-import {
-  MissingParamError,
-  retrieveSecondaryMessage,
-} from "../src/common/error.js";
-import { parseArray, parseBoolean } from "../src/common/ops.js";
-import { renderError } from "../src/common/render.js";
-import { fetchStats } from "../src/fetchers/stats.js";
-import { isLocaleAvailable } from "../src/translations.js";
-
-// @ts-ignore
-export default async ({
-  username,
-  repo,
-  owner,
-  hide,
-  hide_title,
-  hide_border,
-  card_width,
-  hide_rank,
-  show_icons,
-  include_all_commits,
-  commits_year,
-  line_height,
-  title_color,
-  ring_color,
-  icon_color,
-  text_color,
-  text_bold,
-  bg_color,
-  theme,
-  exclude_repo,
-  custom_title,
-  locale,
-  disable_animations,
-  border_radius,
-  number_format,
-  role,
-  number_precision,
-  border_color,
-  rank_icon,
-  show,
-}) => {
-  const access = guardAccess({
-    id: username,
-    type: "username",
-    colors: {
-      title_color,
-      text_color,
-      bg_color,
-      border_color,
-      theme,
-    },
-  });
-  if (!access.isPassed) {
-    return { status: "error - permanent", content: access.result };
-  }
-
-  if (locale && !isLocaleAvailable(locale)) {
-    return {
-      status: "error - permanent",
-      content: renderError({
-        message: "Something went wrong",
-        secondaryMessage: "Language not found",
-        renderOptions: {
-          title_color,
-          text_color,
-          bg_color,
-          border_color,
-          theme,
-        },
-      }),
-    };
-  }
-
-  const safePattern = /^[-\w/.,]+$/;
-  if (
-    (username && !safePattern.test(username)) ||
-    (repo && !safePattern.test(repo)) ||
-    (owner && !safePattern.test(owner))
-  ) {
-    return {
-      status: "error - permanent",
-      content: renderError({
-        message: "Something went wrong",
-        secondaryMessage:
-          "Username, repository or owner contains unsafe characters",
-        renderOptions: {
-          title_color,
-          text_color,
-          bg_color,
-          border_color,
-          theme,
-        },
-      }),
-    };
-  }
-
-  try {
-    const showStats = parseArray(show);
-    const repoOwner = parseArray(owner);
-    let repository = parseArray(repo);
-    repository = repository.map((repo) =>
-      repo.includes("/") ? repo : `${username}/${repo}`,
-    );
-
-    const stats = await fetchStats(
-      username,
-      parseBoolean(include_all_commits),
-      parseArray(exclude_repo),
-      showStats.includes("prs_merged") ||
-        showStats.includes("prs_merged_percentage"),
-      showStats.includes("discussions_started"),
-      showStats.includes("discussions_answered"),
-      parseInt(commits_year, 10),
-      repository,
-      repoOwner,
-      showStats.includes("prs_authored"),
-      showStats.includes("prs_commented"),
-      showStats.includes("prs_reviewed"),
-      showStats.includes("issues_authored"),
-      showStats.includes("issues_commented"),
-      parseArray(role),
-    );
-
-    return {
-      status: "success",
-      content: renderStatsCard(
-        stats,
-        {
-          hide: parseArray(hide),
-          show_icons: parseBoolean(show_icons),
-          hide_title: parseBoolean(hide_title),
-          hide_border: parseBoolean(hide_border),
-          card_width: parseInt(card_width, 10),
-          hide_rank: parseBoolean(hide_rank),
-          include_all_commits: parseBoolean(include_all_commits),
-          commits_year: parseInt(commits_year, 10),
-          line_height,
-          title_color,
-          ring_color,
-          icon_color,
-          text_color,
-          text_bold: parseBoolean(text_bold),
-          bg_color,
-          theme,
-          custom_title,
-          border_radius,
-          border_color,
-          number_format,
-          number_precision: parseInt(number_precision, 10),
-          locale: locale ? locale.toLowerCase() : null,
-          disable_animations: parseBoolean(disable_animations),
-          rank_icon,
-          show: showStats,
-        },
-        username,
-        repository,
-        repoOwner,
-      ),
-    };
-  } catch (err) {
-    if (err instanceof Error) {
-      return {
-        status: "error - temporary",
-        content: renderError({
-          message: err.message,
-          secondaryMessage: retrieveSecondaryMessage(err),
-          renderOptions: {
-            title_color,
-            text_color,
-            bg_color,
-            border_color,
-            theme,
-            show_repo_link: !(err instanceof MissingParamError),
-          },
-        }),
-      };
-    }
-    return {
-      status: "error - temporary",
-      content: renderError({
-        message: "An unknown error occurred",
-        renderOptions: {
-          title_color,
-          text_color,
-          bg_color,
-          border_color,
-          theme,
-        },
-      }),
-    };
-  }
-};
